#7: add plot of dropped nodes; refactor code to be more modular
