+++ detection/utils/Helpers.py
@@ -1,0 +1,3 @@
+def getNodeNumber(node_name: str):
+    number = ''.join(filter(str.isdigit, node_name))
+    return int(number) if number else None
